Refactor LatexSetup and LatexBackref methods out of renderer.py
These are going to get larger and more complicated and the single file was unwieldy.
